Front Blue auto rewrite, removed RobotCommand, AutoStopper, TelMet
