+++ FtcRobotController/src/main/java/org/firstinspires/ftc/robotcontroller/internal/Core/RobotBase.java
@@ -14,14 +14,11 @@ package org.firstinspires.ftc.robotcontroller.internal.Core;
 
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.hardware.HardwareMap;
-
-import org.directcurrent.core.TelMet;
 
 
 /**
- * Base of a robot- every robot base should have these components to run properly. The base acts
+ * Base of a robot- every robot base should have these components to start properly. The base acts
  * as a sort of "tie" between components and sensors- it provides a single object to pass around to
  * OpModes and provides easy hardware mapping.
  */
@@ -30,9 +27,6 @@ public abstract class RobotBase
 {
     /** HardwareMap object used for mapping robot hardware. */
     public HardwareMap hardware = null;
-
-    /** TelMet object for outputting messages to Telemetry */
-    private TelMet _telMet = null;
 
     /** OpMode for goodies such as telemetry and actually being able to stop robot code HAHA */
     private LinearOpMode _opmode = null;
@@ -55,18 +49,7 @@ public abstract class RobotBase
     public void init(final HardwareMap HW , final LinearOpMode OPMODE)
     {
         hardware = HW;
-        _telMet = new TelMet(OPMODE.telemetry);
         _opmode = OPMODE;
-    }
-
-
-    /**
-     * @deprecated Just use opMode().telemetry()
-     * @return Returns the TelMet object for use
-     */
-    public final TelMet telMet()
-    {
-        return _telMet;
     }
