+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/SeasonCode/RelicRecovery/OpModes/Autonomous/AutoBackSideRed.java
@@ -2,17 +2,9 @@ package org.firstinspires.ftc.teamcode.SeasonCode.RelicRecovery.OpModes.Autonomo
 
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.hardware.DcMotor;
 
-import org.directcurrent.core.AutoStopper;
-import org.directcurrent.opencv.CVBridge;
-import org.directcurrent.season.relicrecovery.jewelarm.JewelArm;
 import org.firstinspires.ftc.robotcontroller.internal.Core.Utility.Vuforia;
-import org.firstinspires.ftc.robotcore.external.navigation.RelicRecoveryVuMark;
-import org.firstinspires.ftc.robotcore.external.navigation.VuforiaLocalizer;
 import org.firstinspires.ftc.teamcode.SeasonCode.RelicRecovery.Base;
-import org.firstinspires.ftc.teamcode.SeasonCode.RelicRecovery.Components.GlyphGrabber.GlyphGrabber;
 
 
 /**
