--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/SeasonCode/RelicRecovery/ToggleTelMet.kt
@@ -1,11 +1,0 @@
-@file:Suppress("PackageDirectoryMismatch")
-package org.directcurrent.season.relicrecovery
-
-@JvmField
-val outputDrivetrain = true
-
-@JvmField
-val outputLift = true
-
-@JvmField
-val outputGlyphGrabber = false
