--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/TestCode/ParallelCommandTest/OpParallelCommandTest.java
@@ -1,19 +1,0 @@
-package org.firstinspires.ftc.teamcode.TestCode.ParallelCommandTest;
-
-
-import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-
-
-
-public class OpParallelCommandTest extends LinearOpMode
-{
-    public void runOpMode() throws InterruptedException
-    {
-        waitForStart();
-
-        while(opModeIsActive())
-        {
-            telemetry.addData("Input" , "Left Joystick X: " + gamepad1.left_stick_x);
-        }
-    }
-}
